dependency-injection/src/test/java/learning/spring/dependencyinjection/controllers/PropertyInjectedControllerTest.java
package learning.spring.dependencyinjection.controllers;

import learning.spring.dependencyinjection.controller.PropertyInjectedController;
import learning.spring.dependencyinjection.service.GreetingServiceImpl;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class PropertyInjectedControllerTest {
    private PropertyInjectedController propertyInjectedController;

    @Before
    public void setUp() throws Exception{
        this.propertyInjectedController = new PropertyInjectedController();
        this.propertyInjectedController.greetingServiceImp = new GreetingServiceImpl();
    }

    @Test
    public void testGreeting(){
        assertEquals(GreetingServiceImpl.HELLO, propertyInjectedController.sayHello());
    }
}
